requirements.txt
flask
flask-cors
google-cloud-texttospeech
